add ResourceCard component
