+++ client/src/components/ResourceCard/index.js
@@ -1,0 +1,48 @@
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import Card from '@material-ui/core/Card';
+import CardActionArea from '@material-ui/core/CardActionArea';
+import CardActions from '@material-ui/core/CardActions';
+import CardContent from '@material-ui/core/CardContent';
+import CardMedia from '@material-ui/core/CardMedia';
+import Button from '@material-ui/core/Button';
+import Typography from '@material-ui/core/Typography';
+
+const useStyles = makeStyles({
+    root: {
+        maxWidth: 345,
+    },
+    media: {
+        height: 140,
+    },
+});
+
+export default function MediaCard() {
+    const classes = useStyles();
+
+    return (
+        <Card className={classes.root}>
+            <CardActionArea>
+                <CardMedia
+                    className="media-img"
+                    image="/img/NCADV.jpg"
+                    title="Contemplative Reptile"
+                />
+                <CardContent>
+                    <Typography gutterBottom variant="h5" component="h2">
+                        NCADV
+                    </Typography>
+                    <Typography variant="body2" color="textSecondary" component="p">
+                        Resource for victims and survivors of domestic violence
+                        This includes a list of national crisis organizations and assistance
+                    </Typography>
+                </CardContent>
+            </CardActionArea>
+            <CardActions>
+                <Button size="small" color="primary">
+                    <a href="https://ncadv.org/resources">Learn More</a>
+                </Button>
+            </CardActions>
+        </Card>
+    );
+}
